requirements.txt
attrs
matplotlib
numpy
Pillow